feat: 매퍼 파일 삭제 후 response 파일 내부로 매퍼 함수 옮김. response 값 nullable 하도록 수정
