+++ app/src/main/java/com/eatssu/android/data/dto/response/UserCollegeDepartmentResponse.kt
@@ -1,5 +1,7 @@
 package com.eatssu.android.data.dto.response
 
+import com.eatssu.android.domain.model.College
+import com.eatssu.android.domain.model.Department
 import com.google.gson.annotations.SerializedName
 
 data class UserCollegeDepartmentResponse(
@@ -12,3 +14,15 @@ data class UserCollegeDepartmentResponse(
     @SerializedName("collegeName")
     val collegeName: String?,
 )
+
+fun UserCollegeDepartmentResponse.toDomain(): Pair<College, Department> =
+    Pair(
+        College(
+            collegeId = this.collegeId ?: -1,
+            collegeName = this.collegeName ?: "단과대"
+        ),
+        Department(
+            departmentId = this.departmentId ?: -1,
+            departmentName = this.departmentName ?: "학과"
+        )
+    )
