+++ app/src/main/java/com/eatssu/android/data/repository/UserRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.eatssu.android.data.dto.request.UserDepartmentRequest
 import com.eatssu.android.data.dto.response.BaseResponse
 import com.eatssu.android.data.dto.response.MyNickNameResponse
 import com.eatssu.android.data.dto.response.MyReviewResponse
-import com.eatssu.android.data.mapper.toDomain
+import com.eatssu.android.data.dto.response.toDomain
 import com.eatssu.android.data.service.UserService
 import com.eatssu.android.domain.model.College
 import com.eatssu.android.domain.model.Department
