+++ app/src/main/java/com/eatssu/android/domain/model/College.kt
@@ -1,6 +1,6 @@
 package com.eatssu.android.domain.model
 
 data class College(
-    val collegeId: Int = -1,
-    val collegeName: String = "단과대",
+    val collegeId: Int,
+    val collegeName: String,
 )
